default font and setAppFont() method called from AppFont
I added the ability to set the font for all components in the View directly from the BaseView. The method call is at the top of the constructor under super(APP_TITLE). The font that I selected is not perfect, and we'll have to workshop some other fonts, and even resize some components to accommodate the standard font. But its really easy to swap and try out other fonts